requirements.txt
numpy
h5py
tensorflow-gpu
matplotlib
opencv-python
tqdm
argparse
Pillow
